pupilcard_datavazir/requirements.txt
streamlit
pandas
numpy
plotly
scipy
python-docx
openpyxl
xlsxwriter
statsmodels
kaleido
transformers
torch
python-dotenv
anthropic